Moved channel subscription tests
